+++ test/fixtures/index.js
@@ -7,10 +7,12 @@ module.exports = {
   cardWithMeldLayout: require('./card-with-meld-layout'),
   secondCardWithMeldLayout: require('./second-card-with-meld-layout'),
   cardBackWithMeldLayout: require('./card-back-with-meld-layout'),
+  cardWithMultipleTokens: require('./card-with-multiple-tokens'),
   catalogOfCardNames: require('./catalog-of-card-names'),
   listOfCards: require('./list-of-cards'),
   listOfCardsPage2: require('./list-of-cards-page-2'),
   listOfPrints: require('./list-of-prints'),
   listOfRulings: require('./list-of-rulings'),
-  set: require('./set')
+  set: require('./set'),
+  tokens: require('./tokens')
 }