+++ test/fixtures/tokens.js
@@ -1,0 +1,387 @@
+'use strict'
+
+module.exports = {
+  elephant: {
+    object: 'card',
+    id: '2dbccfc7-427b-41e6-b770-92d73994bf3b',
+    oracle_id: '079c46cc-feb0-4998-8593-c8b739afdb82',
+    multiverse_ids: [
+
+    ],
+    tcgplayer_id: 98663,
+    name: 'Elephant',
+    lang: 'en',
+    released_at: '2015-05-22',
+    uri: 'https://api.scryfall.com/cards/2dbccfc7-427b-41e6-b770-92d73994bf3b',
+    scryfall_uri: 'https://scryfall.com/card/tmm2/9/elephant?utm_source=api',
+    layout: 'token',
+    highres_image: true,
+    image_uris: {
+      small: 'https://img.scryfall.com/cards/small/en/tmm2/9.jpg?1517813031',
+      normal: 'https://img.scryfall.com/cards/normal/en/tmm2/9.jpg?1517813031',
+      large: 'https://img.scryfall.com/cards/large/en/tmm2/9.jpg?1517813031',
+      png: 'https://img.scryfall.com/cards/png/en/tmm2/9.png?1517813031',
+      art_crop: 'https://img.scryfall.com/cards/art_crop/en/tmm2/9.jpg?1517813031',
+      border_crop: 'https://img.scryfall.com/cards/border_crop/en/tmm2/9.jpg?1517813031'
+    },
+    mana_cost: '',
+    cmc: 0.0,
+    type_line: 'Token Creature — Elephant',
+    oracle_text: '',
+    power: '3',
+    toughness: '3',
+    colors: [
+      'G'
+    ],
+    color_identity: [
+      'G'
+    ],
+    all_parts: [
+      {
+        object: 'related_card',
+        id: '73c16134-692d-4fd1-bffa-f9342113cbd8',
+        component: 'combo_piece',
+        name: 'Bestial Menace',
+        type_line: 'Sorcery',
+        uri: 'https://api.scryfall.com/cards/73c16134-692d-4fd1-bffa-f9342113cbd8'
+      },
+      {
+        object: 'related_card',
+        id: '2dbccfc7-427b-41e6-b770-92d73994bf3b',
+        component: 'token',
+        name: 'Elephant',
+        type_line: 'Token Creature — Elephant',
+        uri: 'https://api.scryfall.com/cards/2dbccfc7-427b-41e6-b770-92d73994bf3b'
+      },
+      {
+        object: 'related_card',
+        id: '2a452235-cebd-4e8f-b217-9b55fc1c3830',
+        component: 'token',
+        name: 'Snake',
+        type_line: 'Token Creature — Snake',
+        uri: 'https://api.scryfall.com/cards/2a452235-cebd-4e8f-b217-9b55fc1c3830'
+      },
+      {
+        object: 'related_card',
+        id: '7bdb3368-fee3-4795-a23f-c97555ee7475',
+        component: 'token',
+        name: 'Wolf',
+        type_line: 'Token Creature — Wolf',
+        uri: 'https://api.scryfall.com/cards/7bdb3368-fee3-4795-a23f-c97555ee7475'
+      }
+    ],
+    legalities: {
+      standard: 'not_legal',
+      future: 'not_legal',
+      frontier: 'not_legal',
+      modern: 'not_legal',
+      legacy: 'not_legal',
+      pauper: 'not_legal',
+      vintage: 'not_legal',
+      penny: 'not_legal',
+      commander: 'not_legal',
+      duel: 'not_legal',
+      oldschool: 'not_legal'
+    },
+    games: [
+      'paper'
+    ],
+    reserved: false,
+    foil: false,
+    nonfoil: true,
+    oversized: false,
+    promo: false,
+    reprint: true,
+    set: 'tmm2',
+    set_name: 'Modern Masters 2015 Tokens',
+    set_uri: 'https://api.scryfall.com/sets/f7aa47c6-c1e2-4de5-9a68-4406d84bd6bb',
+    set_search_uri: 'https://api.scryfall.com/cards/search?order=set&q=e%3Atmm2&unique=prints',
+    scryfall_set_uri: 'https://scryfall.com/sets/tmm2?utm_source=api',
+    rulings_uri: 'https://api.scryfall.com/cards/2dbccfc7-427b-41e6-b770-92d73994bf3b/rulings',
+    prints_search_uri: 'https://api.scryfall.com/cards/search?order=released&q=oracleid%3A079c46cc-feb0-4998-8593-c8b739afdb82&unique=prints',
+    collector_number: '9',
+    digital: false,
+    rarity: 'common',
+    illustration_id: '0fe3c48b-f071-458f-9320-fa34b7e5d8cb',
+    artist: 'Lars Grant-West',
+    border_color: 'black',
+    frame: '2015',
+    full_art: false,
+    story_spotlight: false,
+    usd: '0.11',
+    prices: {
+      usd: '0.11',
+      usd_foil: null,
+      eur: null,
+      tix: null
+    },
+    related_uris: {
+      tcgplayer_decks: 'https://decks.tcgplayer.com/magic/deck/search?contains=Elephant&page=1&partner=Scryfall&utm_campaign=affiliate&utm_medium=scryfall&utm_source=scryfall',
+      edhrec: 'http://edhrec.com/route/?cc=Elephant',
+      mtgtop8: 'http://mtgtop8.com/search?MD_check=1&SB_check=1&cards=Elephant'
+    },
+    purchase_uris: {
+      tcgplayer: 'https://shop.tcgplayer.com/magic/modern-masters-2015/elephant-token?partner=Scryfall&utm_campaign=affiliate&utm_medium=scryfall&utm_source=scryfall',
+      cardmarket: 'https://www.cardmarket.com/en/Magic/Products/Search?referrer=scryfall&searchString=Elephant&utm_campaign=card_prices&utm_medium=text&utm_source=scryfall',
+      cardhoarder: 'https://www.cardhoarder.com/cards?affiliate_id=scryfall&data%5Bsearch%5D=Elephant&ref=card-profile&utm_campaign=affiliate&utm_medium=card&utm_source=scryfall'
+    }
+  },
+  wolf: {
+    object: 'card',
+    id: '7bdb3368-fee3-4795-a23f-c97555ee7475',
+    oracle_id: 'b2224843-8274-4872-a7ca-2adf69cc066b',
+    multiverse_ids: [
+
+    ],
+    tcgplayer_id: 98667,
+    name: 'Wolf',
+    lang: 'en',
+    released_at: '2015-05-22',
+    uri: 'https://api.scryfall.com/cards/7bdb3368-fee3-4795-a23f-c97555ee7475',
+    scryfall_uri: 'https://scryfall.com/card/tmm2/13/wolf?utm_source=api',
+    layout: 'token',
+    highres_image: true,
+    image_uris: {
+      small: 'https://img.scryfall.com/cards/small/en/tmm2/13.jpg?1517813031',
+      normal: 'https://img.scryfall.com/cards/normal/en/tmm2/13.jpg?1517813031',
+      large: 'https://img.scryfall.com/cards/large/en/tmm2/13.jpg?1517813031',
+      png: 'https://img.scryfall.com/cards/png/en/tmm2/13.png?1517813031',
+      art_crop: 'https://img.scryfall.com/cards/art_crop/en/tmm2/13.jpg?1517813031',
+      border_crop: 'https://img.scryfall.com/cards/border_crop/en/tmm2/13.jpg?1517813031'
+    },
+    mana_cost: '',
+    cmc: 0.0,
+    type_line: 'Token Creature — Wolf',
+    oracle_text: '',
+    power: '2',
+    toughness: '2',
+    colors: [
+      'G'
+    ],
+    color_identity: [
+      'G'
+    ],
+    all_parts: [
+      {
+        object: 'related_card',
+        id: '73c16134-692d-4fd1-bffa-f9342113cbd8',
+        component: 'combo_piece',
+        name: 'Bestial Menace',
+        type_line: 'Sorcery',
+        uri: 'https://api.scryfall.com/cards/73c16134-692d-4fd1-bffa-f9342113cbd8'
+      },
+      {
+        object: 'related_card',
+        id: 'fb10f5be-a82f-4f4f-8653-2d6927ac0ca8',
+        component: 'combo_piece',
+        name: 'Wolfbriar Elemental',
+        type_line: 'Creature — Elemental',
+        uri: 'https://api.scryfall.com/cards/fb10f5be-a82f-4f4f-8653-2d6927ac0ca8'
+      },
+      {
+        object: 'related_card',
+        id: '2dbccfc7-427b-41e6-b770-92d73994bf3b',
+        component: 'token',
+        name: 'Elephant',
+        type_line: 'Token Creature — Elephant',
+        uri: 'https://api.scryfall.com/cards/2dbccfc7-427b-41e6-b770-92d73994bf3b'
+      },
+      {
+        object: 'related_card',
+        id: '2a452235-cebd-4e8f-b217-9b55fc1c3830',
+        component: 'token',
+        name: 'Snake',
+        type_line: 'Token Creature — Snake',
+        uri: 'https://api.scryfall.com/cards/2a452235-cebd-4e8f-b217-9b55fc1c3830'
+      },
+      {
+        object: 'related_card',
+        id: '7bdb3368-fee3-4795-a23f-c97555ee7475',
+        component: 'token',
+        name: 'Wolf',
+        type_line: 'Token Creature — Wolf',
+        uri: 'https://api.scryfall.com/cards/7bdb3368-fee3-4795-a23f-c97555ee7475'
+      }
+    ],
+    legalities: {
+      standard: 'not_legal',
+      future: 'not_legal',
+      frontier: 'not_legal',
+      modern: 'not_legal',
+      legacy: 'not_legal',
+      pauper: 'not_legal',
+      vintage: 'not_legal',
+      penny: 'not_legal',
+      commander: 'not_legal',
+      duel: 'not_legal',
+      oldschool: 'not_legal'
+    },
+    games: [
+      'paper'
+    ],
+    reserved: false,
+    foil: false,
+    nonfoil: true,
+    oversized: false,
+    promo: false,
+    reprint: true,
+    set: 'tmm2',
+    set_name: 'Modern Masters 2015 Tokens',
+    set_uri: 'https://api.scryfall.com/sets/f7aa47c6-c1e2-4de5-9a68-4406d84bd6bb',
+    set_search_uri: 'https://api.scryfall.com/cards/search?order=set&q=e%3Atmm2&unique=prints',
+    scryfall_set_uri: 'https://scryfall.com/sets/tmm2?utm_source=api',
+    rulings_uri: 'https://api.scryfall.com/cards/7bdb3368-fee3-4795-a23f-c97555ee7475/rulings',
+    prints_search_uri: 'https://api.scryfall.com/cards/search?order=released&q=oracleid%3Ab2224843-8274-4872-a7ca-2adf69cc066b&unique=prints',
+    collector_number: '13',
+    digital: false,
+    rarity: 'common',
+    illustration_id: '5e290bd2-37d0-4eb0-a2fc-fafc7107c42e',
+    artist: 'Daren Bader',
+    border_color: 'black',
+    frame: '2015',
+    full_art: false,
+    story_spotlight: false,
+    usd: '0.20',
+    prices: {
+      usd: '0.20',
+      usd_foil: null,
+      eur: null,
+      tix: null
+    },
+    related_uris: {
+      tcgplayer_decks: 'https://decks.tcgplayer.com/magic/deck/search?contains=Wolf&page=1&partner=Scryfall&utm_campaign=affiliate&utm_medium=scryfall&utm_source=scryfall',
+      edhrec: 'http://edhrec.com/route/?cc=Wolf',
+      mtgtop8: 'http://mtgtop8.com/search?MD_check=1&SB_check=1&cards=Wolf'
+    },
+    purchase_uris: {
+      tcgplayer: 'https://shop.tcgplayer.com/magic/modern-masters-2015/wolf-token?partner=Scryfall&utm_campaign=affiliate&utm_medium=scryfall&utm_source=scryfall',
+      cardmarket: 'https://www.cardmarket.com/en/Magic/Products/Search?referrer=scryfall&searchString=Wolf&utm_campaign=card_prices&utm_medium=text&utm_source=scryfall',
+      cardhoarder: 'https://www.cardhoarder.com/cards?affiliate_id=scryfall&data%5Bsearch%5D=Wolf&ref=card-profile&utm_campaign=affiliate&utm_medium=card&utm_source=scryfall'
+    }
+  },
+  snake: {
+    object: 'card',
+    id: '2a452235-cebd-4e8f-b217-9b55fc1c3830',
+    oracle_id: '9ecd83ce-f866-4cbe-9712-c44aabe979c6',
+    multiverse_ids: [
+
+    ],
+    tcgplayer_id: 98666,
+    name: 'Snake',
+    lang: 'en',
+    released_at: '2015-05-22',
+    uri: 'https://api.scryfall.com/cards/2a452235-cebd-4e8f-b217-9b55fc1c3830',
+    scryfall_uri: 'https://scryfall.com/card/tmm2/12/snake?utm_source=api',
+    layout: 'token',
+    highres_image: true,
+    image_uris: {
+      small: 'https://img.scryfall.com/cards/small/en/tmm2/12.jpg?1517813031',
+      normal: 'https://img.scryfall.com/cards/normal/en/tmm2/12.jpg?1517813031',
+      large: 'https://img.scryfall.com/cards/large/en/tmm2/12.jpg?1517813031',
+      png: 'https://img.scryfall.com/cards/png/en/tmm2/12.png?1517813031',
+      art_crop: 'https://img.scryfall.com/cards/art_crop/en/tmm2/12.jpg?1517813031',
+      border_crop: 'https://img.scryfall.com/cards/border_crop/en/tmm2/12.jpg?1517813031'
+    },
+    mana_cost: '',
+    cmc: 0.0,
+    type_line: 'Token Creature — Snake',
+    oracle_text: '',
+    power: '1',
+    toughness: '1',
+    colors: [
+      'G'
+    ],
+    color_identity: [
+      'G'
+    ],
+    all_parts: [
+      {
+        object: 'related_card',
+        id: '73c16134-692d-4fd1-bffa-f9342113cbd8',
+        component: 'combo_piece',
+        name: 'Bestial Menace',
+        type_line: 'Sorcery',
+        uri: 'https://api.scryfall.com/cards/73c16134-692d-4fd1-bffa-f9342113cbd8'
+      },
+      {
+        object: 'related_card',
+        id: '2dbccfc7-427b-41e6-b770-92d73994bf3b',
+        component: 'token',
+        name: 'Elephant',
+        type_line: 'Token Creature — Elephant',
+        uri: 'https://api.scryfall.com/cards/2dbccfc7-427b-41e6-b770-92d73994bf3b'
+      },
+      {
+        object: 'related_card',
+        id: '2a452235-cebd-4e8f-b217-9b55fc1c3830',
+        component: 'token',
+        name: 'Snake',
+        type_line: 'Token Creature — Snake',
+        uri: 'https://api.scryfall.com/cards/2a452235-cebd-4e8f-b217-9b55fc1c3830'
+      },
+      {
+        object: 'related_card',
+        id: '7bdb3368-fee3-4795-a23f-c97555ee7475',
+        component: 'token',
+        name: 'Wolf',
+        type_line: 'Token Creature — Wolf',
+        uri: 'https://api.scryfall.com/cards/7bdb3368-fee3-4795-a23f-c97555ee7475'
+      }
+    ],
+    legalities: {
+      standard: 'not_legal',
+      future: 'not_legal',
+      frontier: 'not_legal',
+      modern: 'not_legal',
+      legacy: 'not_legal',
+      pauper: 'not_legal',
+      vintage: 'not_legal',
+      penny: 'not_legal',
+      commander: 'not_legal',
+      duel: 'not_legal',
+      oldschool: 'not_legal'
+    },
+    games: [
+      'paper'
+    ],
+    reserved: false,
+    foil: false,
+    nonfoil: true,
+    oversized: false,
+    promo: false,
+    reprint: true,
+    set: 'tmm2',
+    set_name: 'Modern Masters 2015 Tokens',
+    set_uri: 'https://api.scryfall.com/sets/f7aa47c6-c1e2-4de5-9a68-4406d84bd6bb',
+    set_search_uri: 'https://api.scryfall.com/cards/search?order=set&q=e%3Atmm2&unique=prints',
+    scryfall_set_uri: 'https://scryfall.com/sets/tmm2?utm_source=api',
+    rulings_uri: 'https://api.scryfall.com/cards/2a452235-cebd-4e8f-b217-9b55fc1c3830/rulings',
+    prints_search_uri: 'https://api.scryfall.com/cards/search?order=released&q=oracleid%3A9ecd83ce-f866-4cbe-9712-c44aabe979c6&unique=prints',
+    collector_number: '12',
+    digital: false,
+    rarity: 'common',
+    illustration_id: 'fc2707bb-a027-40de-a67d-1a2f7dec3422',
+    artist: 'Austin Hsu',
+    border_color: 'black',
+    frame: '2015',
+    full_art: false,
+    story_spotlight: false,
+    usd: '0.16',
+    prices: {
+      usd: '0.16',
+      usd_foil: null,
+      eur: null,
+      tix: null
+    },
+    related_uris: {
+      tcgplayer_decks: 'https://decks.tcgplayer.com/magic/deck/search?contains=Snake&page=1&partner=Scryfall&utm_campaign=affiliate&utm_medium=scryfall&utm_source=scryfall',
+      edhrec: 'http://edhrec.com/route/?cc=Snake',
+      mtgtop8: 'http://mtgtop8.com/search?MD_check=1&SB_check=1&cards=Snake'
+    },
+    purchase_uris: {
+      tcgplayer: 'https://shop.tcgplayer.com/magic/modern-masters-2015/snake-token?partner=Scryfall&utm_campaign=affiliate&utm_medium=scryfall&utm_source=scryfall',
+      cardmarket: 'https://www.cardmarket.com/en/Magic/Products/Search?referrer=scryfall&searchString=Snake&utm_campaign=card_prices&utm_medium=text&utm_source=scryfall',
+      cardhoarder: 'https://www.cardhoarder.com/cards?affiliate_id=scryfall&data%5Bsearch%5D=Snake&ref=card-profile&utm_campaign=affiliate&utm_medium=card&utm_source=scryfall'
+    }
+  }
+}